+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.2.0",
   "description": "skin without meat",
   "scripts": {
-    "start": "parcel ./src/index.html --port 3000",
+    "start": "node src/srv.js",
     "build": "parcel build ./src/index.html --no-source-maps --target default",
     "clean": "rimraf .parcel-cache && rimraf .cache && rimraf dist && rimraf build"
   },
@@ -11,21 +11,21 @@
   "author": "SSA",
   "license": "ISC",
   "devDependencies": {
-    "@parcel/transformer-inline-string": "^2.6.2",
-    "@types/hbs": "^4.0.1",
-    "@types/node": "^18.6.4",
-    "eslint": "^8.20.0",
-    "events": "^3.3.0",
-    "parcel": "^2.6.2",
-    "postcss": "^8.4.14",
-    "postcss-custom-properties": "^12.1.8",
-    "postcss-import": "^14.1.0",
-    "postcss-url": "^10.1.3",
-    "posthtml-include": "^1.7.4",
-    "prettier": "^2.7.1",
-    "rimraf": "^3.0.2"
+    "@parcel/transformer-inline-string": "2.6.2",
+    "@types/hbs": "4.0.1",
+    "@types/node": "18.6.4",
+    "eslint": "8.20.0",
+    "events": "3.3.0",
+    "parcel": "2.6.2",
+    "postcss": "8.4.14",
+    "postcss-custom-properties": "12.1.8",
+    "postcss-import": "14.1.0",
+    "postcss-url": "10.1.3",
+    "posthtml-include": "1.7.4",
+    "prettier": "2.7.1",
+    "rimraf": "3.0.2"
   },
   "dependencies": {
-    "handlebars": "^4.7.7"
+    "handlebars": "4.7.7"
   }
 }
